refactor(pro-vrf): rename package, classes, methods, variables
