+++ provider-verify/src/main/kotlin/io/github/wliamp/pro/vrf/data/FacebookOauth.kt
@@ -1,16 +1,16 @@
-package io.github.wliamp.provider.data;
+package io.github.wliamp.pro.vrf.data
 
-import io.github.wliamp.provider.config.ProviderProperties
+import io.github.wliamp.pro.vrf.config.VerifyProviderProperties
 import org.springframework.core.ParameterizedTypeReference
 import org.springframework.web.reactive.function.client.WebClient
 import reactor.core.publisher.Mono
 import kotlin.collections.get
 
 class FacebookOauth(
-    private val props: ProviderProperties,
+    private val props: VerifyProviderProperties,
     private val webClient: WebClient
-) : Oauth {
-    private val party = "facebook"
+) : IOauth {
+    private val provider = "facebook"
 
     override fun verify(token: String): Mono<Boolean> =
         if (props.facebookAppId.isBlank() || props.facebookTokenInfoUrl.isBlank() || props.facebookAppAccessToken.isBlank()) Mono.error(
