+++ provider-verify/src/main/kotlin/io/github/wliamp/pro/vrf/data/IOauth.kt
@@ -1,8 +1,8 @@
-package io.github.wliamp.provider.data
+package io.github.wliamp.pro.vrf.data
 
 import reactor.core.publisher.Mono
 
-interface Oauth {
+interface IOauth {
     fun verify(token: String): Mono<Boolean>
     fun getInfo(token: String): Mono<Map<String, Any>>
 }
