+++ provider-verify/src/main/kotlin/io/github/wliamp/pro/vrf/util/OauthProvider.kt
@@ -1,0 +1,13 @@
+package io.github.wliamp.pro.vrf.util
+
+import io.github.wliamp.pro.vrf.data.IOauth
+import org.springframework.stereotype.Component
+
+@Component
+class OauthProvider(
+    val facebook: IOauth,
+    val google: IOauth,
+    val zalo: IOauth
+)
+
+
