--- provider-verify/src/main/kotlin/io/github/wliamp/provider/util/OauthVerifier.kt
@@ -1,13 +1,0 @@
-package io.github.wliamp.provider.util
-
-import io.github.wliamp.provider.data.Oauth
-import org.springframework.stereotype.Component
-
-@Component
-class OauthVerifier(
-    val facebook: Oauth,
-    val google: Oauth,
-    val zalo: Oauth
-)
-
-
