+++ provider-verify/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.autoconfiguration.imports
@@ -1,1 +1,1 @@
-io.github.wliamp.provider.config.ProviderAutoConfig
+io.github.wliamp.pro.vrf.config.VerifyProviderAutoConfig
