+++ provider-verify/src/test/kotlin/io/github/wliamp/pro/vrf/TestClass.kt
@@ -1,4 +1,4 @@
-package io.github.wliamp.provider
+package io.github.wliamp.pro.vrf
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
